Factor out block classes
