+++ Interpret/SmartView/BinaryParser.cpp
@@ -3,13 +3,6 @@
 
 namespace smartview
 {
-	void block::addBlockBytes(const blockBytes& child)
-	{
-		auto _block = child;
-		_block.text = strings::BinToHexString(child, true);
-		children.push_back(_block);
-	}
-
 	static std::wstring CLASS = L"CBinaryParser";
 
 	CBinaryParser::CBinaryParser() { m_Offset = 0; }
