+++ Interpret/SmartView/BinaryParser.h
@@ -1,133 +1,12 @@
 #pragma once
-#include <utility>
+#include <Interpret/SmartView/block/block.h>
+#include <Interpret/SmartView/block/blockT.h>
+#include <Interpret/SmartView/block/blockBytes.h>
+#include <Interpret/SmartView/block/blockStringA.h>
+#include <Interpret/SmartView/block/blockStringW.h>
 
 namespace smartview
 {
-	class blockBytes;
-	class block
-	{
-	public:
-		block() : offset(0), cb(0), text(L""), header(true) {}
-		explicit block(std::wstring _text) : offset(0), cb(0), text(std::move(_text)), header(true) {}
-
-		std::wstring ToString() const
-		{
-			std::vector<std::wstring> items;
-			items.reserve(children.size() + 1);
-			items.push_back(text);
-			for (const auto& item : children)
-			{
-				items.push_back(item.ToString());
-			}
-
-			return strings::join(items, strings::emptystring);
-		}
-		size_t getSize() const { return cb; }
-		void setSize(size_t _size) { cb = _size; }
-		size_t getOffset() const { return offset; }
-		void setOffset(size_t _offset) { offset = _offset; }
-		template <typename... Args> void addHeader(const std::wstring& _text, Args... args)
-		{
-			children.emplace_back(block(strings::formatmessage(_text.c_str(), args...)));
-		}
-		void addBlock(const block& child, const std::wstring& _text)
-		{
-			auto block = child;
-			block.text = _text;
-			children.push_back(block);
-		}
-		template <typename... Args> void addBlock(const block& child, const std::wstring& _text, Args... args)
-		{
-			auto block = child;
-			block.text = strings::formatmessage(_text.c_str(), args...);
-			children.push_back(block);
-		}
-		void addBlock(const block& child) { children.push_back(child); }
-		void addBlockBytes(const blockBytes& child);
-		void addLine() { addHeader(L"\r\n"); }
-		bool hasData() const { return !text.empty() || !children.empty(); }
-
-	protected:
-		size_t offset;
-		size_t cb;
-
-	private:
-		std::wstring text;
-		std::vector<block> children;
-		bool header;
-	};
-
-	class blockBytes : public block
-	{
-	public:
-		void setData(const std::vector<BYTE>& data) { _data = data; }
-		std::vector<BYTE> getData() const { return _data; }
-		operator const std::vector<BYTE>&() const { return _data; }
-		operator const std::vector<BYTE>() const { return _data; }
-		size_t size() const noexcept { return _data.size(); }
-		bool empty() const noexcept { return _data.empty(); }
-		const BYTE* data() const noexcept { return _data.data(); }
-		BYTE operator[](const size_t _Pos) { return _data[_Pos]; }
-
-	private:
-		std::vector<BYTE> _data;
-	};
-
-	template <class T> class blockT : public block
-	{
-	public:
-		blockT() { memset(&data, 0, sizeof(data)); }
-		void setData(const T& _data) { data = _data; }
-		T getData() const { return data; }
-		operator T&() { return data; }
-		operator T() const { return data; }
-		T& operator=(const blockT<T>& _data)
-		{
-			data = _data.data;
-			cb = _data.getSize();
-			offset = _data.getOffset();
-			return *this;
-		}
-		T& operator=(const T& _data)
-		{
-			data = _data;
-			return *this;
-		}
-
-	private:
-		T data;
-	};
-
-	class blockStringA : public block
-	{
-	public:
-		void setData(const std::string& _data) { data = _data; }
-		std::string getData() const { return data; }
-		operator std::string&() { return data; }
-		operator std::string() const { return data; }
-		_Ret_z_ const char* c_str() const noexcept { return data.c_str(); }
-		size_t length() const noexcept { return data.length(); }
-		bool empty() const noexcept { return data.empty(); }
-
-	private:
-		std::string data;
-	};
-
-	class blockStringW : public block
-	{
-	public:
-		void setData(const std::wstring& _data) { data = _data; }
-		std::wstring getData() const { return data; }
-		operator std::wstring&() { return data; }
-		operator std::wstring() const { return data; }
-		_Ret_z_ const wchar_t* c_str() const noexcept { return data.c_str(); }
-		size_t length() const noexcept { return data.length(); }
-		bool empty() const noexcept { return data.empty(); }
-
-	private:
-		std::wstring data;
-	};
-
 	// TODO: Make all of this header only
 	// CBinaryParser - helper class for parsing binary data without
 	// worrying about whether you've run off the end of your buffer.
