+++ MrMapi/MrMAPI.vcxproj
@@ -483,6 +483,11 @@
     <ClInclude Include="..\Interpret\SmartView\AdditionalRenEntryIDs.h" />
     <ClInclude Include="..\Interpret\SmartView\AppointmentRecurrencePattern.h" />
     <ClInclude Include="..\Interpret\SmartView\BinaryParser.h" />
+    <ClInclude Include="..\Interpret\SmartView\block\block.h" />
+    <ClInclude Include="..\Interpret\SmartView\block\blockBytes.h" />
+    <ClInclude Include="..\Interpret\SmartView\block\blockStringA.h" />
+    <ClInclude Include="..\Interpret\SmartView\block\blockStringW.h" />
+    <ClInclude Include="..\Interpret\SmartView\block\blockT.h" />
     <ClInclude Include="..\Interpret\SmartView\ConversationIndex.h" />
     <ClInclude Include="..\Interpret\SmartView\EntryIdStruct.h" />
     <ClInclude Include="..\Interpret\SmartView\EntryList.h" />
@@ -579,6 +584,7 @@
     <ClCompile Include="..\Interpret\SmartView\AdditionalRenEntryIDs.cpp" />
     <ClCompile Include="..\Interpret\SmartView\AppointmentRecurrencePattern.cpp" />
     <ClCompile Include="..\Interpret\SmartView\BinaryParser.cpp" />
+    <ClCompile Include="..\Interpret\SmartView\block\block.cpp" />
     <ClCompile Include="..\Interpret\SmartView\ConversationIndex.cpp" />
     <ClCompile Include="..\Interpret\SmartView\EntryIdStruct.cpp" />
     <ClCompile Include="..\Interpret\SmartView\EntryList.cpp" />
